move helper into plugin/util
